+++ proyectointegrador/client/src/components/ApartadoArtistico.css
@@ -95,34 +95,3 @@
     }
 }
 
-.poemas-grid {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 20px;
-    margin: 0 auto;
-    max-width: 1200px;
-}
-  
-.poema-card {
-    background-color: #ffffff;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); 
-    padding: 20px;
-    text-align: center;
-    transition: transform 0.3s, box-shadow 0.3s;
-} 
-
-.poema-titulo {
-    font-size: 1.5em;
-    color: #3aa871;
-    margin-bottom: 10px;
-}
-
-.boton-visualizacion {
-    border-radius: 10px;
-    background-color: #2e7d32;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    transition: transform 0.3s, box-shadow 0.3s;
-}
-
