+++ proyectointegrador/server/models/config/database.js
@@ -3,7 +3,7 @@ const { Sequelize } = require('sequelize');
 const sequelize = new Sequelize({
   database: "ecochallenge",
   username: "postgres",
-  password: "1",
+  password: "Ale+dra123",
   host: "localhost",
   dialect: "postgres",
 });
